requirements.txt
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.71
botocore==1.40.71
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
mypy==1.18.2
mypy_extensions==1.1.0
pathspec==0.12.1
pydantic==2.12.2
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
